requirements.txt
faiss-cpu
numpy
google-generativeai
python-dotenv
requests
fastapi
uvicorn